rmgmt: wire up storage node to disk space manager
Allows the manager to directly subscribe to disk space alerts.

This change revealed some build dependency issues, and required playing
header/types tetris, and moving some types around so they were
accessible in more locations without causing include loops.
