+++ src/v/resource_mgmt/storage.h
@@ -14,6 +14,7 @@
 #include "config/property.h"
 #include "seastarx.h"
 #include "ssx/semaphore.h"
+#include "storage/node.h"
 
 #include <seastar/core/sharded.hh>
 
@@ -30,27 +31,7 @@ class partition_manager;
 namespace storage {
 
 class api;
-
-enum class disk_space_alert { ok = 0, low_space = 1, degraded = 2 };
-
-inline disk_space_alert max_severity(disk_space_alert a, disk_space_alert b) {
-    return std::max(a, b);
-}
-
-inline std::ostream& operator<<(std::ostream& o, const disk_space_alert d) {
-    switch (d) {
-    case disk_space_alert::ok:
-        o << "ok";
-        break;
-    case disk_space_alert::low_space:
-        o << "low_space";
-        break;
-    case disk_space_alert::degraded:
-        o << "degraded";
-        break;
-    }
-    return o;
-}
+class node;
 
 /*
  *
@@ -60,6 +41,7 @@ public:
     disk_space_manager(
       config::binding<bool> enabled,
       ss::sharded<storage::api>* storage,
+      ss::sharded<storage::node>* storage_node,
       ss::sharded<cloud_storage::cache>* cache,
       ss::sharded<cluster::partition_manager>* pm);
 
@@ -75,6 +57,7 @@ public:
 private:
     config::binding<bool> _enabled;
     ss::sharded<storage::api>* _storage;
+    ss::sharded<storage::node>* _storage_node;
     ss::sharded<cloud_storage::cache>* _cache;
     ss::sharded<cluster::partition_manager>* _pm;
